Altera 'LONG' para 'int32_t' (source) No MS-Windows, o tamanho de LONG é 4 bytes tanto em 32-bit quanto em 64-bit.

+++ source/activex/c_activex.c
@@ -570,7 +570,7 @@ HB_FUNC(HWG_SETUPCONNECTIONPOINT)
 #endif
 
     pThis->pEvents = hb_itemNew(hb_param(3, HB_IT_ANY));
-    hb_stornl((LONG)pThis, 2);
+    hb_stornl((int32_t)pThis, 2);
   }
 
   hb_retnl(hr);
